@@ -4,7 +4,7 @@ version: 0.1.9+1
 homepage: https://github.com/duythien0912/flutter_zalo_login
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: "^1.10.0"
 
 dependencies:
