@@ -18,16 +18,16 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  ZaloProfileModel zaloInfo = ZaloProfileModel(
+  ZaloProfileModel? zaloInfo = ZaloProfileModel(
     birthday: "",
     gender: "",
     id: "",
@@ -35,7 +35,7 @@ class _MyHomePageState extends State<MyHomePage> {
     picture: null,
   );
 
-  ZaloLoginResult zaloLoginResult = ZaloLoginResult(
+  ZaloLoginResult? zaloLoginResult = ZaloLoginResult(
     errorCode: -1,
     errorMessage: "",
     oauthCode: "",
@@ -96,7 +96,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: Text(widget.title!),
       ),
       body: SingleChildScrollView(
         child: Column(
@@ -155,12 +155,12 @@ class _MyHomePageState extends State<MyHomePage> {
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: <Widget>[
                         Text(
-                          "oauthCode: " + zaloLoginResult.oauthCode.toString(),
+                          "oauthCode: ${zaloLoginResult?.oauthCode.toString()}",
                           overflow: TextOverflow.ellipsis,
                           maxLines: 1,
                         ),
-                        Text("errorCode: " + zaloLoginResult.errorCode.toString()),
-                        Text("userId: " + zaloLoginResult.userId),
+                        Text("errorCode: ${ zaloLoginResult?.errorCode.toString()??""}"),
+                        Text("userId: ${zaloLoginResult?.userId!}"),
                       ],
                     ),
                   ),
@@ -211,18 +211,18 @@ class _MyHomePageState extends State<MyHomePage> {
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: <Widget>[
-                        if (zaloInfo != null && zaloInfo.picture != null) Image.network(zaloInfo.picture.data.url),
+                        if (zaloInfo != null && zaloInfo!.picture != null) Image.network(zaloInfo!.picture!.data!.url!),
                         Text(
-                          "id: " + zaloInfo?.id,
+                          "id: ${zaloInfo?.id??""}",
                         ),
                         Text(
-                          "name: " + zaloInfo?.name,
+                          "name: ${zaloInfo?.name??""}",
                         ),
                         Text(
-                          "birthday: " + zaloInfo?.birthday,
+                          "birthday: ${ zaloInfo?.birthday??""}",
                         ),
                         Text(
-                          "gender: " + zaloInfo?.gender,
+                          "gender: ${zaloInfo?.gender??""}",
                         ),
                       ],
                     ),
